+++ tests/unit/sinon-sandbox-with-global-access-test.js
@@ -4,10 +4,6 @@ import { createSandbox, restoreSandbox, setOptions } from 'ember-sinon-sandbox/t
 module('Unit | ember-sinon-sandbox | With global access', {
   before() {
     setOptions({ errorOnGlobalSinonAccess: false });
-  },
-
-  after() {
-    setOptions({ errorOnGlobalSinonAccess: true });
   }
 });
 
@@ -38,87 +34,14 @@ test('stubbing out sandbox.create returns the already created sandbox', function
   restoreSandbox();
 });
 
-test('calling `sinon.sandbox.restore()` can be called explicitly and via `restoreSandbox`', function(assert) {
+test('ensures sandbox is restored correctly', function(assert) {
   assert.expect(2);
 
   createSandbox();
-
-  this.sandbox.spy();
-
-  assert.equal(this.sandbox.fakes.length, 1);
-
-  this.sandbox.restore();
-
-  assert.equal(this.sandbox.fakes.length, 0);
-
-  restoreSandbox();
-});
-
-test('using useFakeTimers API continues to work', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  let clock = this.sandbox.useFakeTimers();
-
-  assert.ok(clock, 'The clock API continues to work after forced sandboxing.');
-
-  restoreSandbox();
-});
-
-test('using useFakeXMLHttpRequest API continues to work', function(assert) {
-  assert.expect(1);
-
-  let requests = [];
-
-  createSandbox();
-
-  let fakeXHR = this.sandbox.useFakeXMLHttpRequest();
-  fakeXHR.onCreate = (req) => {
-    requests.push(req);
-  };
-
-  let xhr = new XMLHttpRequest();
-  xhr.addEventListener("load", () => {});
-  xhr.open("GET", "http://www.example.org/example.txt");
-  xhr.send();
-
-  assert.equal(requests.length, 1, 'The fake XHR API continues to work after forced sandboxing.');
-
-  restoreSandbox();
-});
-
-test('using sinon.assert.* methods throws an error', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  assert.throws(() => {
-    this.sandbox.assert.calledOnce()
-  }, 'sinon.assert methods throw an error');
-
-  restoreSandbox();
-});
-
-test('using sinon.fakeServer.create throws an error', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
-
-  assert.throws(() => {
-    this.sandbox.fakeServer.create()
-  }, 'sandbox.fakeServer.create throws and error');
-
-  restoreSandbox();
-});
-
-test('ensures sandbox is restored correctly', function(assert) {
-  assert.expect(1);
-
-  createSandbox();
   restoreSandbox();
 
-  assert.notOk(self.sinon, 'Sandbox is restored');
+  assert.equal(this.sandbox, null, 'Sandbox is set to null after restoration');
+  assert.equal(self.sinon, null, 'self.sinon is set to null after restoration');
 });
 
 test('ensures sandbox instances are different for each test', function(assert) {
